serving/app/requirements.txt
flask
pillow
google-cloud-storage
requests
numpy